+++ Sources/Danger/Danger.swift
@@ -26,6 +26,12 @@ private final class DangerRunner {
         logger.logInfo("Ran with: \(CommandLine.arguments.joined(separator: " "))")
         
         let cliLength = CommandLine.arguments.count
+        
+        guard cliLength - 2 > 0 else {
+            logger.logError("To execute Danger run danger-swift ci , danger-swift pr or danger-swift local on your terminal")
+            exit(1)
+        }
+        
         let dslJSONArg: String? = CommandLine.arguments[cliLength - 2]
         let outputJSONPath = CommandLine.arguments[cliLength - 1]
 
@@ -65,72 +71,142 @@ public func Danger() -> DangerDSL {
     return DangerRunner.shared.dsl
 }
 
+extension DangerDSL {
+    /// Fails on the Danger report
+    public var fails: [Violation] {
+        return DangerRunner.shared.results.fails
+    }
+    
+    /// Warnings on the Danger report
+    public var warnings: [Violation] {
+        return DangerRunner.shared.results.warnings
+    }
+    
+    /// Messages on the Danger report
+    public var messages: [Violation] {
+        return DangerRunner.shared.results.messages
+    }
+    
+    /// Markdowns on the Danger report
+    public var markdowns: [Violation] {
+        return DangerRunner.shared.results.markdowns
+    }
+    
+    /// Adds a warning message to the Danger report
+    ///
+    /// - Parameter message: A markdown-ish
+    public func warn(_ message: String) {
+        DangerRunner.shared.results.warnings.append(Violation(message: message))
+    }
+    
+    /// Adds an inline warning message to the Danger report
+    public func warn(message: String, file: String, line: Int) {
+        DangerRunner.shared.results.warnings.append(Violation(message: message, file: file, line: line))
+    }
+    
+    /// Adds a warning message to the Danger report
+    ///
+    /// - Parameter message: A markdown-ish
+    public func fail(_ message: String) {
+        DangerRunner.shared.results.fails.append(Violation(message: message))
+    }
+    
+    /// Adds an inline fail message to the Danger report
+    public func fail(message: String, file: String, line: Int) {
+        DangerRunner.shared.results.fails.append(Violation(message: message, file: file, line: line))
+    }
+    
+    /// Adds a warning message to the Danger report
+    ///
+    /// - Parameter message: A markdown-ish
+    public func message(_ message: String) {
+        DangerRunner.shared.results.messages.append(Violation(message: message))
+    }
+    
+    /// Adds an inline message to the Danger report
+    public func message(message: String, file: String, line: Int) {
+        DangerRunner.shared.results.messages.append(Violation(message: message, file: file, line: line))
+    }
+    
+    /// Adds a warning message to the Danger report
+    ///
+    /// - Parameter message: A markdown-ish
+    public func markdown(_ message: String) {
+        DangerRunner.shared.results.markdowns.append(Violation(message: message))
+    }
+    
+    /// Adds an inline message to the Danger report
+    public func markdown(message: String, file: String, line: Int) {
+        DangerRunner.shared.results.markdowns.append(Violation(message: message, file: file, line: line))
+    }
+}
+
 /// Fails on the Danger report
 public var fails: [Violation] {
-    return DangerRunner.shared.results.fails
+    return DangerRunner.shared.dsl.fails
 }
 
 /// Warnings on the Danger report
 public var warnings: [Violation] {
-    return DangerRunner.shared.results.warnings
+    return DangerRunner.shared.dsl.warnings
 }
 
 /// Messages on the Danger report
 public var messages: [Violation] {
-    return DangerRunner.shared.results.messages
+    return DangerRunner.shared.dsl.messages
 }
 
 /// Markdowns on the Danger report
 public var markdowns: [Violation] {
-    return DangerRunner.shared.results.markdowns
+    return DangerRunner.shared.dsl.markdowns
 }
 
 /// Adds a warning message to the Danger report
 ///
 /// - Parameter message: A markdown-ish
 public func warn(_ message: String) {
-    DangerRunner.shared.results.warnings.append(Violation(message: message))
+    DangerRunner.shared.dsl.warn(message)
 }
 
 /// Adds an inline warning message to the Danger report
 public func warn(message: String, file: String, line: Int) {
-    DangerRunner.shared.results.warnings.append(Violation(message: message, file: file, line: line))
+    DangerRunner.shared.dsl.warn(message: message, file: file, line: line)
 }
 
 /// Adds a warning message to the Danger report
 ///
 /// - Parameter message: A markdown-ish
 public func fail(_ message: String) {
-    DangerRunner.shared.results.fails.append(Violation(message: message))
+    DangerRunner.shared.dsl.fail(message)
 }
 
 /// Adds an inline fail message to the Danger report
 public func fail(message: String, file: String, line: Int) {
-    DangerRunner.shared.results.fails.append(Violation(message: message, file: file, line: line))
+    DangerRunner.shared.dsl.fail(message: message, file: file, line: line)
 }
 
 /// Adds a warning message to the Danger report
 ///
 /// - Parameter message: A markdown-ish
 public func message(_ message: String) {
-    DangerRunner.shared.results.messages.append(Violation(message: message))
+    DangerRunner.shared.dsl.message(message)
 }
 
 /// Adds an inline message to the Danger report
 public func message(message: String, file: String, line: Int) {
-    DangerRunner.shared.results.messages.append(Violation(message: message, file: file, line: line))
+    DangerRunner.shared.dsl.message(message: message, file: file, line: line)
 }
 
 /// Adds a warning message to the Danger report
 ///
 /// - Parameter message: A markdown-ish
 public func markdown(_ message: String) {
-    DangerRunner.shared.results.markdowns.append(Violation(message: message))
+    DangerRunner.shared.dsl.markdown(message)
 }
 
 /// Adds an inline message to the Danger report
 public func markdown(message: String, file: String, line: Int) {
-    DangerRunner.shared.results.markdowns.append(Violation(message: message, file: file, line: line))
+    DangerRunner.shared.dsl.markdown(message: message, file: file, line: line)
 }
 
 // MARK: - Private Functions